Add ondeck battery overlays and limp-home mode

+++ bb/src/main/java/com/richardmcdougall/bb/hardware/VESC_CAN_STATUS_MSG.java
@@ -2,7 +2,7 @@ package com.richardmcdougall.bb.hardware;
 
 public class VESC_CAN_STATUS_MSG {
     int id;
-    int rx_time;
+    long rx_time;
     float rpm;
     float current;
     float duty;
