+++ bb/src/main/java/com/richardmcdougall/bb/hardware/VESC_CAN_STATUS_MSG_3.java
@@ -2,7 +2,7 @@ package com.richardmcdougall.bb.hardware;
 
 public class VESC_CAN_STATUS_MSG_3 {
     int id;
-    int rx_time;
+    long rx_time;
     float watt_hours;
     float watt_hours_charged;
