+++ bb/src/main/java/com/richardmcdougall/bb/hardware/VESC_CAN_STATUS_MSG_5.java
@@ -4,7 +4,7 @@ package com.richardmcdougall.bb.hardware;
 
 public class VESC_CAN_STATUS_MSG_5 {
     int id;
-    int rx_time;
+    long rx_time;
     float v_in;
     int tacho_value;
