+++ bb/src/main/java/com/richardmcdougall/bb/hardware/VESC_CAN_STATUS_MSG_6.java
@@ -4,7 +4,7 @@ package com.richardmcdougall.bb.hardware;
 
 public class VESC_CAN_STATUS_MSG_6 {
     int id;
-    int rx_time;
+    long rx_time;
     float adc_1;
     float adc_2;
     float adc_3;
